* add json_schema_utils * BaseDataInterface uses base schema, not root * moved special root json schemas from utils to nwbconverter * subset_channels is defined in the init of BaseRecordingExtractorInterface * corresponding changes to subset_channels for Neuroscope * input validation moved to top of function
