+++ nwb_conversion_tools/basedatainterface.py
@@ -2,27 +2,27 @@
 from abc import abstractmethod
 import warnings
 
-from .utils import get_base_source_schema, get_base_metadata_schema, get_schema_from_method_signature
+from .json_schema_utils import get_base_schema, get_schema_from_method_signature
 
 
 class BaseDataInterface:
 
     @classmethod
     def get_source_schema(cls):
-        return get_base_source_schema()
+        return get_base_schema()
 
     def __init__(self, **source_data):
         self.source_data = source_data
 
     def get_metadata_schema(self):
-        return get_base_metadata_schema()
+        return get_base_schema()
 
     def get_metadata(self):
         return dict()
 
     @classmethod
     def get_conversion_options_schema(cls):
-        return get_schema_from_method_signature(cls.convert_data, exclude=['nwbfile', 'metadata'])
+        return get_schema_from_method_signature(cls.run_conversion, exclude=['nwbfile', 'metadata'])
 
     @abstractmethod
     def run_conversion(self, nwbfile_path, metadata, **conversion_options):
