+++ nwb_conversion_tools/baserecordingextractorinterface.py
@@ -4,8 +4,8 @@ from pynwb.device import Device
 from pynwb.ecephys import ElectrodeGroup, ElectricalSeries
 
 from .basedatainterface import BaseDataInterface
-from .utils import (get_schema_from_method_signature, get_schema_from_hdmf_class,
-                    get_metadata_schema)
+from .utils import get_schema_from_hdmf_class
+from .json_schema_utils import get_schema_from_method_signature
 
 
 class BaseRecordingExtractorInterface(BaseDataInterface):
@@ -18,10 +18,10 @@ class BaseRecordingExtractorInterface(BaseDataInterface):
     def __init__(self, **source_data):
         super().__init__(**source_data)
         self.recording_extractor = self.RX(**source_data)
+        self.subset_channels = None
 
     def get_metadata_schema(self):
-        # Initiate empty metadata schema
-        metadata_schema = get_metadata_schema()
+        metadata_schema = super().get_metadata_schema()
 
         # Initiate Ecephys metadata
         metadata_schema['properties']['Ecephys'] = dict(
@@ -55,32 +55,28 @@ class BaseRecordingExtractorInterface(BaseDataInterface):
 
         Parameters
         ----------
-        nwbfile : NWBFile object
-        metadata_dict : dictionary
-        stub_test : boolean, optional (default False)
+        nwbfile: NWBFile object
+        metadata: dictionary
+        stub_test: boolean, optional (default False)
             If True, will truncate the data to run the conversion faster and take up less memory.
         """
-        if stub_test:
-            num_frames = 100
-            test_ids = self.recording_extractor.get_channel_ids()
-            end_frame = min([num_frames, self.recording_extractor.get_num_frames()])
 
-            stub_recording_extractor = se.SubRecordingExtractor(
-                self.recording_extractor,
-                channel_ids=test_ids,
-                start_frame=0,
-                end_frame=end_frame
-            )
-        else:
-            stub_recording_extractor = self.recording_extractor
+        recording_extractor = self.recording_extractor
+        if stub_test or self.subset_channels is not None:
+            kwargs = dict()
 
-        if metadata is not None and 'Ecephys' in metadata and 'subset_channels' in metadata['Ecephys']:
+            if stub_test:
+                num_frames = 100
+                end_frame = min([num_frames, self.recording_extractor.get_num_frames()])
+                kwargs.update(end_frame=end_frame)
+
+            if self.subset_channels is not None:
+                kwargs.update(channel_ids=self.subset_channels)
+
             recording_extractor = se.SubRecordingExtractor(
-                stub_recording_extractor,
-                channel_ids=metadata['Ecephys']['subset_channels']
+                self.recording_extractor,
+                **kwargs
             )
-        else:
-            recording_extractor = stub_recording_extractor
 
         se.NwbRecordingExtractor.write_recording(
             recording_extractor,
