+++ nwb_conversion_tools/nwbconverter.py
@@ -6,9 +6,8 @@ from jsonschema import validate
 from pynwb import NWBHDF5IO, NWBFile
 from pynwb.file import Subject
 
-from .utils import (get_schema_from_hdmf_class, get_base_metadata_schema,
-                    get_base_source_schema, get_base_conversion_options_schema,
-                    get_schema_for_NWBFile, dict_deep_update)
+from .utils import get_schema_from_hdmf_class, get_schema_for_NWBFile
+from .json_schema_utils import dict_deep_update, get_root_schema
 
 
 class NWBConverter:
@@ -19,17 +18,29 @@ class NWBConverter:
     @classmethod
     def get_source_schema(cls):
         """Compile input schemas from each of the data interface classes."""
-        source_schema = get_base_source_schema()
+        source_schema = get_root_schema()
+        source_schema.update({
+            "$id": "source.schema.json",
+            "title": "Source data schema",
+            "description": "Schema for the source data, files and directories",
+            "version": "0.1.0",
+        })
         for name, data_interface in cls.data_interface_classes.items():
             source_schema['properties'].update(name=data_interface.get_source_schema())
         return source_schema
 
     @classmethod
     def get_conversion_options_schema(cls):
-        conversion_options = get_base_conversion_options_schema()
+        conversion_options_schema = get_root_schema()
+        conversion_options_schema.update({
+            "$id": "conversion_options.schema.json",
+            "title": "Conversion options schema",
+            "description": "Schema for the conversion options",
+            "version": "0.1.0",
+        })
         for name, data_interface in cls.data_interface_classes.items():
-            conversion_options['properties'].update(name=data_interface.get_conversion_options_schema())
-        return conversion_options
+            conversion_options_schema['properties'].update(name=data_interface.get_conversion_options_schema())
+        return conversion_options_schema
 
     def __init__(self, **source_data):
         """
@@ -39,19 +50,26 @@ class NWBConverter:
         validate(instance=source_data, schema=self.get_source_schema())
 
         # If data is valid, proceed to instantiate DataInterface objects
-        self.data_interface_objects = {name: data_interface(**source_data[name])
-                                       for name, data_interface in
-                                       self.data_interface_classes.items()}
+        self.data_interface_objects = {
+            name: data_interface(**source_data[name])
+            for name, data_interface in self.data_interface_classes.items()
+        }
 
     def get_metadata_schema(self):
         """Compile metadata schemas from each of the data interface objects."""
-        metadata_schema = get_base_metadata_schema()
-        metadata_schema['properties'] = dict(
-            NWBFile=get_schema_for_NWBFile(),
-            Subject=get_schema_from_hdmf_class(Subject)
-        )
-        metadata_schema['required'].append('NWBFile')
-        for name, data_interface in self.data_interface_objects.items():
+        metadata_schema = get_root_schema()
+        metadata_schema.update({
+            "$id": "metadata.schema.json",
+            "title": "Metadata",
+            "description": "Schema for the metadata",
+            "version": "0.1.0",
+            "required": ["NWBFile"],
+            "properties": dict(
+                NWBFile=get_schema_for_NWBFile(),
+                Subject=get_schema_from_hdmf_class(Subject)
+            )
+        })
+        for data_interface in self.data_interface_objects.values():
             interface_schema = data_interface.get_metadata_schema()
             metadata_schema = dict_deep_update(metadata_schema, interface_schema)
         return metadata_schema
@@ -62,7 +80,7 @@ class NWBConverter:
             NWBFile=dict(
                 session_description="no description",
                 identifier=str(uuid.uuid4()),
-                session_start_time=datetime.strptime('1900-01-01 00:00:00', '%Y-%m-%d %H:%M:%S'),
+                session_start_time=datetime(1900, 1, 1)
             )
         )
         for interface in self.data_interface_objects.values():
@@ -70,13 +88,15 @@ class NWBConverter:
             metadata = dict_deep_update(metadata, interface_metadata)
         return metadata
 
-    def run_conversion(self, metadata, nwbfile_path=None, save_to_file=True, conversion_options=None):
+    def run_conversion(self, metadata: dict, nwbfile_path=None, save_to_file: bool = True,
+                       conversion_options: dict = None):
         """Build nwbfile object, auto-populate with minimal values if missing."""
         if conversion_options is None:
             conversion_options = dict()
-        nwbfile_kwargs = dict()
-        if 'NWBFile' in metadata:
-            nwbfile_kwargs.update(metadata['NWBFile'])
+        else:
+            validate(instance=conversion_options, schema=self.get_conversion_options_schema())
+
+        nwbfile_kwargs = metadata['NWBFile']
         if 'Subject' in metadata:
             nwbfile_kwargs.update(subject=Subject(**metadata['Subject']))
         # convert ISO 8601 string to datetime
@@ -85,9 +105,6 @@ class NWBConverter:
                 metadata['NWBFile']['session_start_time']
             )
         nwbfile = NWBFile(**nwbfile_kwargs)
-
-        # Validate conversion_options against conversion_options_schema
-        validate(instance=conversion_options, schema=self.get_conversion_options_schema())
 
         # If conversion_options is valid, procede to run conversions for DataInterfaces
         for interface_name, data_interface in self.data_interface_objects.items():
